Project_Foodity/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
mysqlclient>=2.2,<3.0
Pillow>=10.0,<11.0
cloudinary>=1.40,<2.0
google-generativeai>=0.5,<1.0
sentry-sdk>=1.40,<3.0
djangorestframework-simplejwt>=5.3,<6.0
